prefect_setup/requirements.txt
pandas
sqlalchemy
psycopg2-binary
prefect
prefect-sqlalchemy
python-dotenv
kaggle